aku melakukan optimasi di feedback python dengan mengirim nilai bilangan bulat dan pembulatan, kemudian aku melakukan pengubahan satuan g/dL ke g/L, aku melakukan perbaikan di display oled, mengubah 2 digit ke 3 digit pada bagian hr,spo2,hb, aku juga melakukan optimasi atau antisipasi pada layar oled ketika terjadi nilai 1 digit agar tetap presisi dan enak dipandang

+++ firmware_arduino/include/Global.h
@@ -33,12 +33,12 @@ struct DataSensor {
 };
 
 struct HasilVitals {
-      float hr;
-      float spo2;
-      float sbp;
-      float dbp;
-      float hb;
-      int std;
+      uint8_t hr;
+      uint8_t spo2;
+      uint8_t sbp;
+      uint8_t dbp;
+      uint8_t hb;
+      uint16_t std;
 };
 
 enum SystemState {
